### What this PR does

This change removes the creation of `StorageNodeUncommitReport`, which was
unnecessary and added complexity. Its cost was mitigated by `sync.Pool`, but it
is no longer needed. The code now directly uses slices of
`LogStreamUncommitReport` for improved simplicity and maintainability.

This change improves performance significantly, reducing memory allocations by
100% in some cases and execution time by up to 75%, as shown below:

```txt
goos: darwin
goarch: arm64
pkg: github.com/kakao/varlog/internal/metarepos
cpu: Apple M3 Max
                            │   old.txt    │               new.txt               │
                            │    sec/op    │   sec/op     vs base                │
SwapReport/getAndSavePtr-14    44.95n ± 1%   10.90n ± 0%  -75.75% (p=0.000 n=20)
SwapReport/getAndSave-14       45.60n ± 0%   11.30n ± 0%  -75.21% (p=0.000 n=20)
SwapReport/swap-14            13.780n ± 0%   8.535n ± 0%  -38.06% (p=0.000 n=20)
geomean                        30.45n        10.17n       -66.61%

                            │   old.txt    │                 new.txt                 │
                            │     B/op     │    B/op     vs base                     │
SwapReport/getAndSavePtr-14   32.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=20)
SwapReport/getAndSave-14      32.00 ± 0%      0.00 ± 0%  -100.00% (p=0.000 n=20)
SwapReport/swap-14            0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=20) ¹
geomean                                  ²               ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean

                            │   old.txt    │                 new.txt                 │
                            │  allocs/op   │ allocs/op   vs base                     │
SwapReport/getAndSavePtr-14   1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=20)
SwapReport/getAndSave-14      1.000 ± 0%     0.000 ± 0%  -100.00% (p=0.000 n=20)
SwapReport/swap-14            0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=20) ¹
geomean                                  ²               ?                       ² ³
¹ all samples are equal
² summaries must be >0 to compute geomean
³ ratios must be >0 to compute geomean
```